+++ Mindstorms.Core/Commands/Button/PressAndReleaseButton.cs
@@ -11,16 +11,16 @@ namespace Mindstorms.Core.Commands.Button
             data = DirectCommandNoReply;
             data.AddRange(new byte[]
             {
-                (byte)OpCode.Button,
-                (byte)ButtonEvent.Press,
-                (byte)button,
+                OpCode.Button,
+                ButtonEvent.Press,
+                button,
 
-                (byte)OpCode.Button,
-                (byte)ButtonEvent.WaitForPress,
+                OpCode.Button,
+                ButtonEvent.WaitForPress,
 
-                (byte)OpCode.Button,
-                (byte)ButtonEvent.Release,
-                (byte)button
+                OpCode.Button,
+                ButtonEvent.Release,
+                button
             });
         }
     }
