+++ Mindstorms.Core/Commands/File/ContinueUploadFileToBrick.cs
@@ -8,7 +8,7 @@ namespace Mindstorms.Core.Commands.File
         public ContinueUploadFileToBrick(byte fileHandle, IEnumerable<byte> fileContent)
         {
             data = SystemCommandNoReply;
-            data.Add((byte)SystemCommand.ContinueFileDownload);
+            data.Add(SystemCommand.ContinueFileDownload);
             data.Add(fileHandle);
             data.AddRange(fileContent);
         }
