+++ Mindstorms.Core/Commands/LCD/DrawCircle.cs
@@ -8,9 +8,9 @@ namespace Mindstorms.Core.Commands.LCD
         public DrawCircle(byte x, byte y, byte radius, LCDColor color, bool fill)
         {
             data = DirectCommandNoReply;
-            data.Add((byte)OpCode.DrawUI);
-            data.Add((byte)(fill ? DrawSubCode.FillCircle : DrawSubCode.Circle));
-            data.AppendOneBytesParameter((byte)color);
+            data.Add(OpCode.DrawUI);
+            data.Add(fill ? DrawSubCode.FillCircle : DrawSubCode.Circle);
+            data.AppendOneBytesParameter(color);
             data.AppendTwoBytesParameter(x);
             data.AppendTwoBytesParameter(y);
             data.AppendTwoBytesParameter(radius);
