+++ Mindstorms.Core/Commands/LCD/DrawLine.cs
@@ -8,9 +8,9 @@ namespace Mindstorms.Core.Commands.LCD
         public DrawLine(byte x1, byte y1, byte x2, byte y2, LCDColor color)
         {
             data = DirectCommandNoReply;
-            data.Add((byte)OpCode.DrawUI);
-            data.Add((byte)DrawSubCode.Line);
-            data.AppendOneBytesParameter((byte)color);
+            data.Add(OpCode.DrawUI);
+            data.Add(DrawSubCode.Line);
+            data.AppendOneBytesParameter(color);
             data.AppendOneBytesParameter(x1);
             data.AppendOneBytesParameter(y1);
             data.AppendOneBytesParameter(x2);
