+++ Mindstorms.Core/Commands/LCD/DrawRectangle.cs
@@ -8,9 +8,9 @@ namespace Mindstorms.Core.Commands.LCD
         public DrawRectangle(byte x, byte y, byte width, byte height, LCDColor color, bool fill)
         {
             data = DirectCommandNoReply;
-            data.Add((byte)OpCode.DrawUI);
-            data.Add((byte)(fill ? DrawSubCode.FillRectangle : DrawSubCode.Rectangle));
-            data.AppendOneBytesParameter((byte)color);
+            data.Add(OpCode.DrawUI);
+            data.Add(fill ? DrawSubCode.FillRectangle : DrawSubCode.Rectangle);
+            data.AppendOneBytesParameter(color);
             data.AppendTwoBytesParameter(x);
             data.AppendTwoBytesParameter(y);
             data.AppendTwoBytesParameter(width);
