+++ Mindstorms.Core/Commands/LCD/Restore.cs
@@ -9,8 +9,8 @@ namespace Mindstorms.Core.Commands.LCD
         public Restore(byte level)
         {
             data = DirectCommandNoReply;
-            data.Add((byte)OpCode.DrawUI);
-            data.Add((byte)DrawSubCode.Restore);
+            data.Add(OpCode.DrawUI);
+            data.Add(DrawSubCode.Restore);
             data.AppendOneBytesParameter(level);
         }
     }
